Fix performance issues caused by 8d15a25dcbb23
The earlier commit would maintain a static list of styles, and every time
a style resource was added to a theme, it would stay there until the end
of time (or termination of JVM). When executing thousands of tests, this
would lead to massive slowdown of test execution - over 2x increase in
our company's case.

Now, the duplicate styles are removed and the new one is added, which keeps
the list short and performance reasonable.

Change-Id: Ib7ed9321c31938577c78aab52a1d8380bf1be4e6