Implement View animations through shadows.
Real android executes these animations through the call to RootViewImpl.setView() but this invokes an out of process binder calls to WindowSession.addToDisplay(...) and requires the return value to be meaningful, plus some other conditions also need to be right.

https://github.com/android/platform_frameworks_base/blob/master/core/java/android/view/ViewRootImpl.java#L527

Somehow then the following calls are invoked by the Android system:-

 RootViewImpl.scheduleTraversals() -> TraversalRunnable() -> doTraversal() -> performTraversals() that eventually calls onDraw() which the calls ViewGroup.draw() and eventually ViewGroup.drawChild() which performs the animations.

The other more complicated solution would be to try and mimic what is happening on real android when we set up the Activity through the ActivityController, we might be able to set the state correctly so that the animations happen but I tried it for a while and it is much more involved.